order_bot/requirements.txt
aiosqlite3
python-dotenv
aiogram==3.0.0b9